Add a basic logic to the RadarController
This commit refactors the `/radar` endpoint so that it can return the
position and targets for the first radar sent to the server in the
payload of a POST request. Currently the spec fails as the attack mode
must determine which radar record and which targets are rendered in the
JSON response.