Fix attack mode bug
The attack mode was hard-coded in a previous commit meaning that only
the `ClosestFirst` mode was ever processed. This commit refactors the
`#radar_params` method defined in the RadarControllerso that the attack
mode is permitted correctly. The mode can now be used in the
`AttackMode.call` method.

For now, only the first attack mode will be used by the AttackModes
modules. However in a later commit, I will need to handle requests with
multiple modes.