Handle different different Attack Modes target order
A later commit will implement the PrioritizeTX Attack Mode which changes
the target order in the API response. As such, it should be possible to
pass in the target order as an argument in the Radar model's
serialisation method. Because complicated SQL statements are expected,
the method will now make use of Arel queries to order a list of targets.