Move logic for the target order into AttackMode
This commit moves some of the logic that would have been used to
determine the order in which targets are displayed into the AttackMode
class. I've created a `.target_order` method in PrioritizeTX class in
particular as so far this seems to be the only Attack Mode that
requires a specific target order.