Add AttackModes::Base class
This commit adds a new Base class for the AttackModes so as to simplify
the addition of new AttackModes in the future. This was also done in an
attempt to accommodate chained Attack Modes however, it's unclear how
these should be treated as the BASH script contains few examples of
these. I'd originally interpreted the spec as requiring a change of
target order in cased like these and added a test for this case. However
on a later reading of the spec, I became unsure as to whether this
change was even required (hence its current pending status.)