Remove comments in boilerplate code
Please note that I'm only removing this because I have been docked
points for not doing this when applying for roles in the past. Ordinary,
I would ask the tech lead or refer to the style guide in use before
making changes like this as I'm sure different teams will have different
preferences.